requirement.txt
absl-py==1.3.0
actionlib==1.14.0
addict==2.4.0
angles==1.9.13
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==23.7.0
bleach==5.0.1
bondpy==1.8.6
brotlipy==0.7.0
cachetools==5.2.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
cmake==3.25.2
colored-logs==0.2.10
coloredlogs==15.0.1
comm==0.1.4
ConfigArgParse==1.5.3
contourpy==1.0.6
controller-manager==0.19.6
controller-manager-msgs==0.19.6
CppHeaderParser==2.7.3.1
cryptography==37.0.2
cv-bridge==1.16.2
cycler==0.11.0
Cython==0.29.33
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
data==0.4
debugpy==1.6.4
decorator==4.0.2
defusedxml==0.7.1
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dpcpp-cpp-rt==2023.0.0
dynamic-reconfigure==1.7.3
easydict==1.10
einops==0.6.0
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.13.1
Flask==2.2.2
fonttools==4.38.0
fsspec==2023.6.0
funcsigs==1.0.2
future==0.18.3
fvcore==0.1.5.post20221221
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.60.0
grpcio==1.51.1
guppy3==3.1.2
httplib2==0.22.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
image-geometry==1.16.2
imgui==1.4.1
impi-rt==2021.8.0
importlib-metadata==7.0.0
importlib-resources==5.10.0
inflection==0.5.1
iniconfig==1.1.1
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-fortran-rt==2023.0.0
intel-numpy==1.21.4
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
intel-scipy==1.7.3
interactive-markers==1.12.0
iopath==0.1.9
ipdb==0.13.11
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.10
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.7
jupyter-console==6.4.4
jupyter_core==5.1.0
jupyter_packaging==0.12.3
jupyter-server==1.23.3
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
jupyterlab-widgets==3.0.3
keras==2.8.0
kiwisolver==1.4.4
laser_geometry==1.6.7
latex==0.7.0
libclang==16.0.6
lit==16.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
memory-profiler==0.61.0
message-filters==1.16.0
mistune==2.0.4
mkl==2023.0.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mkl-umath==0.1.1
mmcv-full==1.4.8
mmdet==2.19.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.0.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
numpy-quaternion==2022.4.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
open3d==0.16.0
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
packaging==22.0
pandas==1.3.1
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.3.0
pip==23.3
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.11.0
pluggy==1.0.0
ply==3.11
plyfile==0.7.4
point-cloud-utils==0.29.1
portalocker==2.7.0
progressbar2==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py3-ready==0.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycocotools==2.0.6
pycparser==2.21
PyDrive==1.3.1
pygcransac==0.1.0
pyglet==1.5.27
Pygments==2.13.0
PyMaxflow==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-qt-binding==0.4.4
python-utils==3.8.1
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.4.0
QtPy==2.2.1
Quaternion==3.5.2.post4
requests==2.28.2
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rviz==1.14.20
safetensors==0.3.1
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sensor-msgs==1.13.1
setuptools==60.2.0
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0.post1
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.12
tabulate==0.9.0
tbb==2021.8.0
tbb4py==2021.8.0
tempdir==0.7.1
tenacity==8.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.8.4
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.17.0
terminaltables==3.1.10
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==3.1.0
timm==0.5.4
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
--find-links https://download.pytorch.org/whl/torch_stable.html
topic-tools==1.16.0
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchsummary==1.5.1
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.6.0
trimesh==3.17.0
triton==2.1.0
typing_extensions==4.7.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.3
wrapt==1.12.1
xacro==1.14.16
yacs==0.1.8
yapf==0.40.2
zipp==3.11.0
imageio==2.5
tyro==0.3.38
wandb==0.13.9
transformers==4.26.1
faiss-gpu==1.7.2
fast-pytorch-kmeans
albumentations==1.3.1
